refactor: combine use cases together and restructure the project
