+++ data/src/main/java/com/london/data/di/RepositoryModule.kt
@@ -17,7 +17,6 @@ import com.london.data.remote.source.authentication.AuthenticationRemoteDataSour
 import com.london.data.remote.source.details.actor.ActorDetailsRemoteDataSource
 import com.london.data.remote.source.details.movie.MovieDetailsRemoteDataSource
 import com.london.data.remote.source.details.tvshow.TvShowDetailsRemoteDataSource
-import com.london.data.remote.source.details.videoprovider.movie.MovieVideoProviderRemote
 import com.london.data.remote.source.details.videoprovider.tvshow.TvShowVideoProviderRemote
 import com.london.data.remote.source.discover.DiscoverRemoteDataSource
 import com.london.data.remote.source.home.popular.PopularRemoteDataSource
@@ -26,13 +25,11 @@ import com.london.data.remote.source.home.trending.TrendingRemoteDataSourceImpl
 import com.london.data.remote.source.home.upcoming.UpComingRemoteDataSource
 import com.london.data.remote.source.reviews.ReviewsRemoteDataSource
 import com.london.data.remote.source.search.SearchRemoteDataSource
-import com.london.data.remote.source.toprated.movie.TopRatedMovieRemoteDataSource
-import com.london.data.remote.source.toprated.tvseries.TopRatedTvRemoteDataSource
+import com.london.data.remote.source.toprated.TopRatedRemoteDataSource
 import com.london.data.repository.authentication.AuthenticationRepositoryImpl
 import com.london.data.repository.discover.DiscoverRepositoryImpl
 import com.london.data.repository.home.popular.PopularRepositoryImpl
-import com.london.data.repository.home.toprated.TopRatedMovieRepositoryImpl
-import com.london.data.repository.home.toprated.TopRatedTvSeriesRepositoryImpl
+import com.london.data.repository.home.toprated.TopRatedRepositoryImpl
 import com.london.data.repository.home.trending.TrendingRepositoryImpl
 import com.london.data.repository.home.upcoming.UpComingRepositoryImpl
 import com.london.data.repository.recent.RecentSearchRepositoryImpl
@@ -40,7 +37,6 @@ import com.london.data.repository.recent.RecentViewedRepositoryImpl
 import com.london.data.repository.recent.RecentWatchedRepositoryIml
 import com.london.data.repository.search.ActorRepositoryImpl
 import com.london.data.repository.search.MovieDetailsRepositoryImpl
-import com.london.data.repository.search.MovieVideoProviderRepositoryImpl
 import com.london.data.repository.search.SearchRepositoryImpl
 import com.london.data.repository.search.TvShowRepositoryImpl
 import com.london.data.repository.search.TvShowVideoProviderRepositoryImpl
@@ -51,7 +47,6 @@ import com.london.domain.entity.recent.RecentViewed
 import com.london.domain.repository.ActorRepository
 import com.london.domain.repository.AuthRepository
 import com.london.domain.repository.MovieDetailsRepository
-import com.london.domain.repository.MovieVideoProviderRepository
 import com.london.domain.repository.PopularRepository
 import com.london.domain.repository.RecentRepository
 import com.london.domain.repository.RecentWatchedRepository
@@ -61,8 +56,7 @@ import com.london.domain.repository.TvShowRepository
 import com.london.domain.repository.TvShowVideoProviderRepository
 import com.london.domain.repository.UpComingRepository
 import com.london.domain.repository.discover.DiscoverRepository
-import com.london.domain.repository.toprated.TopRatedMovieRepository
-import com.london.domain.repository.toprated.TopRatedTvSeriesRepository
+import com.london.domain.repository.toprated.TopRatedRepository
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -119,27 +113,14 @@ object RepositoryModule {
 
     @Provides
     @Singleton
-    fun provideTopRatedMovieRepository(
-        dataSource: TopRatedMovieRemoteDataSource,
+    fun provideTopRatedRepository(
+        dataSource: TopRatedRemoteDataSource,
         @Named("topRatedLocalDataSource") localTopRated: HomeLocalDataSource<TopRatedLocal>,
         crashReporter: CrashReporter
-    ): TopRatedMovieRepository =
-        TopRatedMovieRepositoryImpl(
-            topRatedMovieRemoteDataSource = dataSource,
+    ): TopRatedRepository =
+        TopRatedRepositoryImpl(
+            topRatedRemoteDataSource = dataSource,
             localTopRated = localTopRated,
-            crashReporter = crashReporter
-        )
-
-    @Provides
-    @Singleton
-    fun provideTopRatedTvSeriesRepository(
-        dataSource: TopRatedTvRemoteDataSource,
-        @Named("topRatedLocalDataSource") localTopRated: HomeLocalDataSource<TopRatedLocal>,
-        crashReporter: CrashReporter
-    ): TopRatedTvSeriesRepository =
-        TopRatedTvSeriesRepositoryImpl(
-            topRatedTvRemoteDataSource = dataSource,
-            topRatedTvShow = localTopRated,
             crashReporter = crashReporter
         )
 
@@ -184,12 +165,6 @@ object RepositoryModule {
 
     @Provides
     @Singleton
-    fun provideMovieVideoRepository(
-        dataSource: MovieVideoProviderRemote
-    ): MovieVideoProviderRepository = MovieVideoProviderRepositoryImpl(dataSource)
-
-    @Provides
-    @Singleton
     fun provideSearchRepository(
         genreInterestDao: GenreInterestDao,
         remoteDataSource: SearchRemoteDataSource,
@@ -226,5 +201,4 @@ object RepositoryModule {
     ): DiscoverRepository = DiscoverRepositoryImpl(
         remoteDataSource = dataSource
     )
-
 }
