--- data/src/main/java/com/london/data/remote/service/toprated/TopRatedMovieApiService.kt
@@ -1,17 +1,0 @@
-@file:KoverIgnore
-
-package com.london.data.remote.service.toprated
-
-import com.london.data.remote.model.ApiResponse
-import com.london.data.remote.model.home.toprated.TopRatedMovieRemote
-import com.london.domain.KoverIgnore
-import retrofit2.Response
-import retrofit2.http.GET
-import retrofit2.http.Query
-
-interface TopRatedMovieApiService {
-    @GET("3/movie/top_rated")
-    suspend fun getTopRatedMovies(
-        @Query("page") pageNumber: Int,
-    ): Response<ApiResponse<TopRatedMovieRemote>>
-}
