+++ data/src/main/java/com/london/data/remote/source/toprated/TopRatedRemoteDataSource.kt
@@ -1,13 +1,19 @@
 @file:KoverIgnore
 
-package com.london.data.remote.source.toprated.movie
+package com.london.data.remote.source.toprated
 
 import com.london.data.remote.model.ApiResponse
 import com.london.data.remote.model.home.toprated.TopRatedMovieRemote
+import com.london.data.remote.model.home.toprated.TopRatedTvSeriesRemote
 import com.london.domain.KoverIgnore
 
-interface TopRatedMovieRemoteDataSource {
+interface TopRatedRemoteDataSource {
+
     suspend fun getTopRatedMovies(
         pageNumber: Int,
     ): Result<ApiResponse<TopRatedMovieRemote>>
+
+    suspend fun getTopRatedTvShows(
+        pageNumber: Int,
+    ): Result<ApiResponse<TopRatedTvSeriesRemote>>
 }
