--- data/src/main/java/com/london/data/remote/source/toprated/movie/TopRatedMovieRemoteDataSourceImpl.kt
@@ -1,22 +1,0 @@
-@file:KoverIgnore
-
-package com.london.data.remote.source.toprated.movie
-
-import com.london.data.remote.model.ApiResponse
-import com.london.data.remote.model.home.toprated.TopRatedMovieRemote
-import com.london.data.remote.service.toprated.TopRatedMovieApiService
-import com.london.data.remote.source.base.BaseRemoteDatasource
-import com.london.domain.KoverIgnore
-import javax.inject.Inject
-
-class TopRatedMovieRemoteDataSourceImpl @Inject constructor(
-    private val topRatedMovieApi: TopRatedMovieApiService
-) : TopRatedMovieRemoteDataSource , BaseRemoteDatasource {
-    override suspend fun getTopRatedMovies(
-        pageNumber: Int,
-    ): Result<ApiResponse<TopRatedMovieRemote>> =
-        callApiWithRetry(
-            apiCall = { topRatedMovieApi.getTopRatedMovies(pageNumber) },
-            mapper = { it }
-        )
-}
