--- data/src/main/java/com/london/data/remote/source/toprated/tvseries/TopRatedTvRemoteDataSource.kt
@@ -1,13 +1,0 @@
-@file:KoverIgnore
-
-package com.london.data.remote.source.toprated.tvseries
-
-import com.london.data.remote.model.ApiResponse
-import com.london.data.remote.model.home.toprated.TopRatedTvSeriesRemote
-import com.london.domain.KoverIgnore
-
-interface TopRatedTvRemoteDataSource {
-    suspend fun getTopRatedTvShows(
-        pageNumber: Int,
-    ): Result<ApiResponse<TopRatedTvSeriesRemote>>
-}
