+++ data/src/main/java/com/london/data/repository/search/MovieDetailsRepositoryImpl.kt
@@ -3,6 +3,7 @@ package com.london.data.repository.search
 import com.london.data.mapper.details.movie.toEntity
 import com.london.data.mapper.search.toEntity
 import com.london.data.mapper.search.toReviewEntity
+import com.london.data.mapper.videoprovider.movie.toMovie
 
 import com.london.data.remote.source.details.movie.MovieDetailsRemoteDataSource
 import com.london.data.remote.source.reviews.ReviewsRemoteDataSource
@@ -14,6 +15,7 @@ import com.london.domain.entity.Movie
 import com.london.domain.entity.PagedFetchResponse
 import com.london.domain.entity.moviedatails.MovieDetails
 import com.london.domain.entity.review.ReviewEntity
+import com.london.domain.entity.videoprovider.MovieVideo
 import com.london.domain.repository.MovieDetailsRepository
 import javax.inject.Inject
 
@@ -68,5 +70,11 @@ class MovieDetailsRepositoryImpl @Inject constructor(
             totalItems = totalItems
         )
     }
+
+    override suspend fun getMovieVideos(movieId: Int): List<MovieVideo> =
+        movieDetailsRemoteDataSource.getMovieVideos(movieId)
+            .getOrThrow().movies?.map { movieVideoRemote ->
+                movieVideoRemote.toMovie()
+            }.orEmpty()
 
 }
