--- data/src/main/java/com/london/data/repository/search/MovieVideoProviderRepositoryImpl.kt
@@ -1,17 +1,0 @@
-package com.london.data.repository.search
-
-import com.london.data.mapper.videoprovider.movie.toMovie
-import com.london.data.remote.source.details.videoprovider.movie.MovieVideoProviderRemote
-import com.london.domain.entity.videoprovider.MovieVideo
-import com.london.domain.repository.MovieVideoProviderRepository
-import javax.inject.Inject
-
-class MovieVideoProviderRepositoryImpl @Inject constructor(
-    private val movieVideoProviderRemote: MovieVideoProviderRemote,
-) : MovieVideoProviderRepository {
-    override suspend fun getMovieVideos(movieId: Int): List<MovieVideo> =
-        movieVideoProviderRemote.getMovieVideos(movieId)
-            .getOrThrow().movies?.map { movieVideoRemote ->
-            movieVideoRemote.toMovie()
-        }.orEmpty()
-}
