+++ data/src/test/java/com/london/data/repository/MovieVideoProviderRepositoryImplTest.kt
@@ -4,9 +4,7 @@ import com.google.common.truth.Truth.assertThat
 import com.london.data.remote.model.details.videoprovider.movie.model.MovieVideoRemote
 import com.london.data.remote.model.details.videoprovider.movie.model.MovieVideoRemoteResponse
 import com.london.data.remote.source.details.videoprovider.movie.MovieVideoProviderRemote
-import com.london.data.repository.search.MovieVideoProviderRepositoryImpl
 import com.london.domain.entity.videoprovider.MovieVideo
-import com.london.domain.repository.MovieVideoProviderRepository
 import io.mockk.coEvery
 import io.mockk.mockk
 import kotlinx.coroutines.test.runTest
