+++ data/src/test/java/com/london/data/repository/TopRatedTvSeriesRepositoryImplTest.kt
@@ -7,7 +7,6 @@ import com.london.data.mapper.home.toprated.toEntity
 import com.london.data.remote.model.ApiResponse
 import com.london.data.remote.model.home.toprated.TopRatedTvSeriesRemote
 import com.london.data.remote.source.toprated.tvseries.TopRatedTvRemoteDataSource
-import com.london.data.repository.home.toprated.TopRatedTvSeriesRepositoryImpl
 import com.london.data.utils.CrashReporter
 import com.london.domain.entity.PagedFetchResponse
 import com.london.domain.entity.toprated.TopRatedTvSeries
