+++ domain/src/main/java/com/london/domain/repository/MovieDetailsRepository.kt
@@ -5,6 +5,7 @@ import com.london.domain.entity.Movie
 import com.london.domain.entity.PagedFetchResponse
 import com.london.domain.entity.moviedatails.MovieDetails
 import com.london.domain.entity.review.ReviewEntity
+import com.london.domain.entity.videoprovider.MovieVideo
 
 
 interface MovieDetailsRepository {
@@ -13,4 +14,5 @@ interface MovieDetailsRepository {
     suspend fun getMovieImagesById(id: Int): List<String>
     suspend fun getMovieCastById(id: Int): List<Actor>
     suspend fun getMovieReviews(movieId: Int, pageNumber: Int): PagedFetchResponse<ReviewEntity>
+    suspend fun getMovieVideos(movieId: Int): List<MovieVideo>
 }
