--- domain/src/main/java/com/london/domain/repository/MovieVideoProviderRepository.kt
@@ -1,7 +1,0 @@
-package com.london.domain.repository
-
-import com.london.domain.entity.videoprovider.MovieVideo
-
-interface MovieVideoProviderRepository {
-    suspend fun getMovieVideos(movieId: Int): List<MovieVideo>
-}
