--- domain/src/main/java/com/london/domain/repository/toprated/TopRatedTvSeriesRepository.kt
@@ -1,10 +1,0 @@
-package com.london.domain.repository.toprated
-
-import com.london.domain.entity.PagedFetchResponse
-import com.london.domain.entity.toprated.TopRatedTvSeries
-
-interface TopRatedTvSeriesRepository {
-    suspend fun getTopRatedTvSeries(
-        pageNumber: Int,
-    ): PagedFetchResponse<TopRatedTvSeries>
-}
