+++ presentation/src/main/java/com/london/presentation/feature/details/movieDetalis/MovieDetailsViewModel.kt
@@ -4,11 +4,7 @@ import androidx.lifecycle.SavedStateHandle
 import com.london.domain.entity.Movie
 import com.london.domain.entity.recent.MediaType
 import com.london.domain.entity.recent.RecentViewed
-import com.london.domain.usecase.details.movie.GetFirstTenMovieImagesUseCase
-import com.london.domain.usecase.details.movie.GetMovieCastUseCase
-import com.london.domain.usecase.details.movie.GetMovieDetailsById
-import com.london.domain.usecase.details.movie.GetMovieVideoUseCase
-import com.london.domain.usecase.details.movie.GetSimilarMoviesUseCase
+import com.london.domain.usecase.details.movie.ManageMovieDetailsUseCase
 import com.london.domain.usecase.recent.viewed.AddToRecentViewedUseCase
 import com.london.domain.usecase.recent.watched.AddMovieToRecentWatchedUseCase
 import com.london.presentation.feature.base.BaseViewModel
@@ -19,11 +15,7 @@ import javax.inject.Inject
 
 @HiltViewModel
 class MovieDetailsViewModel @Inject constructor(
-    private val getMovieById: GetMovieDetailsById,
-    private val getMovieImagesUseCase: GetFirstTenMovieImagesUseCase,
-    private val getMovieCastUseCase: GetMovieCastUseCase,
-    private val getSimilarMoviesUseCase: GetSimilarMoviesUseCase,
-    private val getMovieVideosUseCase: GetMovieVideoUseCase,
+    private val movieDetails: ManageMovieDetailsUseCase,
     private val addMovieToRecentWatchedUseCase:AddMovieToRecentWatchedUseCase,
     private val addToRecentViewedUseCase:AddToRecentViewedUseCase,
     savedStateHandle: SavedStateHandle
@@ -67,10 +59,10 @@ class MovieDetailsViewModel @Inject constructor(
     private fun loadMovieDetails(movieId: Int) {
         tryToExecute(
             block = {
-                val movieDetails = getMovieById.invoke(movieId)
-                val movieImages = getMovieImagesUseCase.invoke(movieId)
-                val movieCast = getMovieCastUseCase.invoke(movieId)
-                Triple(movieDetails, movieImages, movieCast)
+                val movie = movieDetails.getMovieDetails(movieId)
+                val movieImages = movieDetails.getFirstTenMovieImagesUseCase(movieId)
+                val movieCast = movieDetails.getMovieCast(movieId)
+                Triple(movie, movieImages, movieCast)
             },
             onStart = { updateState { copy(isLoading = true) } },
             onSuccess = { triple ->
@@ -133,8 +125,8 @@ class MovieDetailsViewModel @Inject constructor(
     private fun loadSimilarAndVideos(movieId: Int) {
         tryToExecute(
             block = {
-                val similarMovies = getSimilarMoviesUseCase.invoke(movieId)
-                val movieVideos = getMovieVideosUseCase.invoke(movieId)
+                val similarMovies = movieDetails.getSimilarMovies(movieId)
+                val movieVideos = movieDetails.getMovieVideo(movieId)
                 Pair(similarMovies, movieVideos)
             },
             onSuccess = { pair ->
